feat: add a better brute force solver
This adds the faster `brute2` solver.

The `brute2` solver will soon replace the old `brute` solver and
this is why `brute2` does not have a better name.

The reason for having both `brute2` and `brute` is for
showing how `brute2` was tested, showing that it returns the same
solutions as `brute` (see `brute2_test.go`).

The basis of `brute2` is testing combinations. This has much
better complexity than the rather silly approach based on
permutations used by `brute`.