chore: check for duplicate subsets in MakeInstance
Disallow duplicate/identical subsets for simplicity.

An optimal solution should always use the cheapest duplicate. But
duplicate subsets can cause some technical issues with fractional
solutions and branching, depending on the implementation. Instead
of doing the book keeping to remove them inside the solver and then
mapping the solution found back to the user's input, we disallow
them.

`TestCheaperSolutionFound` in `brute_test.go` was changed because
it used an instance that had duplicate subsets.