chore: make BenchmarkCheckSubsetsForDuplicates fairer
This should make the benchmark less biased.

Before the subsets were sorted by construction in the benchmark
which made `checkSubsetsForDuplicates` perform better
than expected. This is because `checkSubsetsForDuplicates` as-is
sorts the subsets as the first step. Other possible methods
could use hash tables instead and wouldn't be sensitive the sorting
of the input.