perf: mostly adopt slog for logging
This is for performance. The adoption of the structured part of `slog`
was done hastily.

The package `slog` has logging levels so we can have debug logging in `algo.go`
and `bb.go` that we can turn off. Now for when running
`BenchmarkBBOnRandomInstances` the logging goes from 35% to not showing up in
the profiling.

The classic `log` only remains in
`cmd/generate_test_instances_and_solutions/main.go`
since it uses `log.Panic(err)` which is not easily replaced.