fix: fix branch indices
This fixes the following bug:
Running the B&B solver on the cost scale=1000 instances introduced in
the previous commit revealed a bug in the branching logic. This bug
resulted in an infinite loop with same branch indicies chosen repeatly.
This meant the sub-instance of node and its parents were the identical.